requirements.txt
annotated-types==0.7.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
requests==2.32.3
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
